python-kafka/requirements.txt
Flask~=3.0.2
flask_restful==0.3.10
kafka-python == 2.0.2